requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
argon2-cffi=20.1.0=py37h7b6447c_1
astroid=2.4.2=py37_0
async_generator=1.10=py37h28b3542_0
attrs=20.2.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
ca-certificates=2020.10.14=0
certifi=2020.6.20=py37_0
cffi=1.14.3=py37he30daa8_0
chardet=3.0.4=pypi_0
cmake=3.18.2.post1=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
easydict=1.9=pypi_0
entrypoints=0.3=py37_0
faiss-cpu=1.6.3=py37h6bb024c_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
insightface=0.1.5=pypi_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.18.1=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
isort=5.6.4=py_0
jedi=0.17.2=py37_0
jinja2=2.11.2=py_0
joblib=0.17.0=pypi_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py37_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.2.0=pypi_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
lxml=4.6.0=pypi_0
markupsafe=1.1.1=py37h14c3975_1
matplotlib=3.3.2=pypi_0
mccabe=0.6.1=pypi_0
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
mxnet-cu102mkl=1.6.0.post0=pypi_0
mxnet-mkl=1.6.0=pypi_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37_0
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=py_0
networkx=2.5=pypi_0
notebook=6.1.4=py37_0
numpy=1.19.2=pypi_0
numpy-base=1.19.1=py37hfa32c7d_0
opencv-python=4.4.0.44=pypi_0
openssl=1.1.1h=h7b6447c_0
packaging=20.4=py_0
pandas=1.1.3=py37he6710b0_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
pexpect=4.8.0=py37_1
pickleshare=0.7.5=py37_1001
pillow=8.0.0=pypi_0
pip=20.2.4=pypi_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
psycopg2-binary=2.8.6=pypi_0
ptyprocess=0.6.0=py37_0
pycparser=2.20=py_2
pygments=2.7.1=py_0
pylint=2.6.0=py37_0
pyparsing=2.4.7=py_0
pypika=0.42.1=pypi_0
pyrsistent=0.17.3=py37h7b6447c_0
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python-graphviz=0.8.4=pypi_0
pytz=2020.1=py_0
pywavelets=1.1.1=pypi_0
pyzmq=19.0.2=py37he6710b0_1
readline=8.0=h7b6447c_0
requests=2.24.0=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.2=pypi_0
send2trash=1.5.0=py37_0
setuptools=50.3.0=py37hb0f4dca_1
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
terminado=0.9.1=py37_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tifffile=2020.10.1=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
tornado=6.0.4=py37h7b6447c_1
tqdm=4.50.2=pypi_0
traitlets=5.0.5=py_0
typed-ast=1.4.1=py37h7b6447c_0
urllib3=1.25.10=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
wheel=0.35.1=py_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.3.0=py_0
zlib=1.2.11=h7b6447c_3
